requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cmake==3.24.0
Django==4.1
djangorestframework==3.13.1
#face-recognition==1.3.0
#face-recognition-models==0.3.0
idna==3.3
numpy==1.23.1
Pillow==9.2.0
pip==21.3.1
PyMySQL==1.0.2
pytz==2022.1
requests==2.28.1
setuptools==60.2.0
sqlparse==0.4.2
urllib3==1.26.11
wheel==0.37.1
